requirements.txt
aenum==2.2.3
gremlinpython==3.4.6
isodate==0.6.0
marshmallow==3.6.0
mlspeclib==0.0.9
PyMySQL==0.9.3
python-box==4.2.3
PyYAML==5.3.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
semver==2.10.1
six==1.14.0
toml==0.10.1
tornado==5.1.1
uritools==3.0.0
